requirements.txt
Django==4.2.6
django-rest-framework==0.1.0